Add div with class image-side and div with class form-side
Reset all margin and padding in web page with all selector.

Set body display to flex, width to 100dvw, height to 100dvh.

Set flex for div with class image-side and div with class form-side to fill window browser with same part.

Set background div with class image-side for searching the assets directory and cover half of the web page.